# perf(rust, python): SIMD accelerated `arg_min`/`arg_max` (via `argminmax`)