--- a/polars/polars-ops/Cargo.toml
+++ b/polars/polars-ops/Cargo.toml
@@ -10,12 +10,12 @@ description = "More operations on polars data structures"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+argminmax = { git = "https://github.com/jvdd/argminmax" }
 arrow.workspace = true
 base64 = { version = "0.21", optional = true }
 either.workspace = true
 hex = { version = "0.4", optional = true }
 jsonpath_lib = { version = "0.3.0", optional = true, git = "https://github.com/ritchie46/jsonpath", branch = "improve_compiled" }
-argminmax = { git = "https://github.com/jvdd/argminmax" }
 memchr.workspace = true
 polars-arrow = { version = "0.28.0", path = "../polars-arrow", default-features = false }
 polars-core = { version = "0.28.0", path = "../polars-core", features = ["private"], default-features = false }