@ritchie46 2023-04-08
Thanks @jvdd that are some impressive benchmarks. :raised_hands: 

I have left some comments.