@jvdd 2023-04-11
If you can point me to where add some more extensive tests (either in Python or Rust) - I’ll gladly include some in this PR :)