@jvdd 2023-04-14
I've added the `"nightly_simd"` as a (default) feature to argminmax since most `x86`/`x86_64` SIMD intrinsics have been stabilized in Rust (except for `AVX512`). Leaving this compute power (`SSE` & `AVX(2)` unused would have been a pity :wink: 

With this feature, the unstable  `AVX512` and `NEON` SIMD are toggled. I've created a new release that includes this feature and updated the code to disable it by default. When the user enables the polars `"simd"` feature, this `"nightly_simd"` feature will be enabled.