@jvdd 2023-04-14
> > With this feature, the unstable AVX512 and NEON SIMD are toggled. I've created a new release that includes this feature and updated the code to disable it by default. When the user enables the polars "simd" feature, this "nightly_simd" feature will be enabled.
> 
> That's great. And on `stable`? It will default to a fallback?

Yes!